Cria a estrutura inicial do projeto instaldo router
