+++ 10_MINIBLOG/miniblog/package.json
@@ -9,6 +9,7 @@
     "firebase": "^9.9.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
