+++ 10_MINIBLOG/miniblog/src/App.js
@@ -1,9 +1,24 @@
+// Imports do router:
+import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
+
+//Styles:
 import './App.css';
+import About from "./pages/About/About";
+
+//Pages:
+import Home from "./pages/Home/Home";
 
 function App() {
   return (
-    <div className="App">      
-      <h1>Miniblog</h1>    
+    <div className="App">
+      <BrowserRouter>
+        <div className="container">
+          <Routes>
+            <Route path='/' element={<Home />} />
+            <Route path='/about' element={<About />} />
+          </Routes>
+        </div>
+      </BrowserRouter>
     </div>
   );
 }
