+++ 10_MINIBLOG/miniblog/src/pages/About/About.js
@@ -1,0 +1,11 @@
+
+//CSS
+import styles from "./About.module.css"
+
+const About = () => {
+  return (
+    <div>About</div>
+  )
+}
+
+export default About
