+++ 10_MINIBLOG/miniblog/src/pages/Home/Home.js
@@ -1,0 +1,11 @@
+
+//CSS
+import styles from "./Home.module.css"
+
+const Home = () => {
+  return (
+    <div>Home</div>
+  )
+}
+
+export default Home
